+++ test/hexapdf/type/test_font_simple.rb
@@ -13,13 +13,17 @@ describe HexaPDF::Type::FontSimple do
       <22> <0042>
       endbfchar
     EOF
-    font_descriptor = @doc.add({Type: :FontDescriptor, FontName: :Embedded, Flags: 0b100,
-                                FontBBox: [0, 1, 2, 3], ItalicAngle: 0, Ascent: 900,
-                                MissingWidth: 500, Descent: -100, CapHeight: 800, StemV: 20})
+    @font_descriptor = @doc.add({Type: :FontDescriptor, FontName: :Embedded, Flags: 0b100,
+                                 FontBBox: [0, 1, 2, 3], ItalicAngle: 0, Ascent: 900,
+                                 MissingWidth: 500, Descent: -100, CapHeight: 800, StemV: 20})
     @font = @doc.add({Type: :Font, Encoding: :WinAnsiEncoding,
-                      BaseFont: :Embedded, FontDescriptor: font_descriptor, ToUnicode: cmap,
+                      BaseFont: :Embedded, FontDescriptor: @font_descriptor, ToUnicode: cmap,
                       FirstChar: 32, LastChar: 34, Widths: [600, 0, 700]},
                      type: HexaPDF::Type::FontSimple)
+  end
+
+  it "returns the font descriptor of the font" do
+    assert_same(@font_descriptor, @font.font_descriptor)
   end
 
   describe "encoding" do